ghu_browser: Don't define global nav element width
Allow templates that inherit from layout.html to set the width of nav
elements. This is a little messy, but I think it's better than defining
it in a separate file from where the nav elements are defined, since if
we add more in the future, we'll be more likely to know to update the
width.