generate_key: If .ini already has [secrets], exit
Currently, if you call generate_key.py on a config.ini which already has
a [secrets] section, it will append another, causing a cryptic error
when you try to `runserver`. So if the .ini exists, parse it to check
that it doesn't already have a [secrets] section before adding one.

Fixes issue #53.