about: Rework design for compactness, WYSIWYG-friendliness
Some of the text was pretty big and stretched across the screen, so make
it more compact. (This is a personal preference thing)

Rework the markup of about.html such that the text in #container can be
written easily a WYSIWYG. Ideally, we can convert about.css into a
page.css and about.html into page.html, which a `page' view calls,
passing in the HTML page contents.

Getting the about-specific CSS into the page is harder, but I think we
could either (1) add a `css' field to Page for direct CSS entry; (2)
same, except with a CSS file name relative to static_root; or (3) we
abstract the CSS away, like have a `this is special page X' field and
another model which maps these special pages to static resources to
include. Definitely something to discuss.