checkbox_blanker: Monitor checkbox with onchange
Listen for the `onchange' event for the 'Make homepage' checkbox, not
the `oninput' event. According to the spec[1] supposedly, oninput only
fires for changes to the input attribute, but checking/unchecking the
'Make homepage' checkbox actually changes only the checked attribute.

This happened in Chrome but not Firefox. Thanks to Sun for finding this!

1: https://developer.mozilla.org/en-US/docs/Web/Events/input