deploy: Allow the ansible script to target a certain branch
Allow passing an argument to ansible-playbook to tell ghu.yml to push
changes to config.ini (for example) only to either master or develop. By
default, it does both. A change to config.ini that broke master but
develop needed prompted this feature. Without it, you end up having to
ssh in by hand, which is annoying and nearly defeats the point of the
ansible script.